Finalization of swipe to search

+++ app/src/main/java/com/fan/tiptop/citiapi/data/CitibikeStationInformationModelDecorated.kt
@@ -3,6 +3,6 @@ package com.fan.tiptop.citiapi.data
 class CitibikeStationInformationModelDecorated(
     val model: CitibikeStationInformationModel,
     val isFavorite: Boolean,
-    var distanceRank: Float= Float.MAX_VALUE
+    var distanceRank: Double= Double.MAX_VALUE
 ) {
 }
